LPS-196807 Use the new method to get the entryClassPK from the base model
Calculates correctly the entryClassPK by delegating the calculation to the asset renderer factory. Also, consider that the asset entry could not exist, for example, in the case of published versions that are not the latest ones.

The entry class PK of the resource models does not always match the resourcePrimKey. There could be more than one asset entry from the same article, which happens in the case of non-visible versions (draft, scheduled, pending...).

In this scenario, only the last approved version, or the first one in case no published version, will use the article resourcePrimKey as her asset entryClassPK. The other non-visible versions will use its ID as entryClassPK.