LPD-23361 Rename PageVersion to PageSpecification.
Rationale: We don't fully support versioning for Pages yet (for example the name of the page cannot be modified only within the draft of a page). Therefore to avoid misunderstandings, and in case we need the term PageVersion in the future if page versioning is implemented, we are going to use PageSpecification instead.

A SitePage can have a draft PageSpecification, a published PageSpecification or both at the same time.