LPD-42808 Adding unit test to reproduce the issue
Jackson cannot infer the type discriminator if the class is not a direct
instance of the polymorphic subclass. When we instantiate the subclass
with an initializer block we are actually creating a subtype of the
subclass which causes the serialization to fail.

This CURL reproduces the issue:

curl -X 'GET' \
  'http://localhost:8080/o/headless-admin-site/v1.0/sites/L_GUEST/site-pages' \
  -H 'accept: application/json' \
  -u 'test@liferay.com:test'

Returning this JSON:

{
    ...
    "pageSettings" : {
      "type" : "SitePageDTOConverter$2",
      "hiddenFromNavigation" : true,
      "layoutTemplateId" : "1_2_columns_i",
      "type" : "WidgetPageSettings"
    },
    ...
}