LPD-XXX  WIP - Attempt to find a fix - Playwright flaky test failing when creating a Template for a Widget
The reason the test fails is that the "Save" button is clicked "too early" (check https://playwright.dev/docs/navigations#hydration for details)

It can be reproduced manually by opening Chrome DevTools,picking "3G" network emulation in the Network tab, entering a name for the template and clicking "Save" before the page finishes loading.

The template will not be saved, which is what is happening during the playwright test, and the reason the test fails at a later step, since the expected template is missing, so it cannot be assigned.

(Got help from Georgel trying to find a fix for it, but not yet working)